change ForbiddenPair to take in UInt8 instead of Int64

+++ ZDD/zdd.jl
@@ -4,28 +4,28 @@ import Base: isequal, ==
 
 # comp₁ < comp₂ by requirement
 struct ForbiddenPair
-    comp₁::Int
-    comp₂::Int
+    comp₁::UInt8
+    comp₂::UInt8
 end
 
 # The 1st node is always the 0-terminal, and the 2nd node is always the 1 terminal. Adding the first node to the ZDD
 # means the ZDD will have 3 nodes, the node + the two terminal nodes
 mutable struct Node
     label::LightGraphs.SimpleGraphs.SimpleEdge{Int64}
-    comp::Array{Int8, 1}
+    comp::Array{UInt8, 1} # can hold 256 possible values if TODO: make UInt8
     comp_weights::Dict{Int8, Int8}
-    cc::Int8
+    cc::UInt8 # can hold only 256 possible values if TODO: make UInt8
     fps::Set{ForbiddenPair}
-    comp_assign::Vector{Int8}
+    comp_assign::Vector{UInt8} # only 256 possible values if TODO: make UInt8
 end
 
 function Node(i::Int)::Node # for Terminal Nodes
-    return Node(Edge(i, i), Array{Int8, 1}(),Dict{Int8, Int8}(), 0, Set{ForbiddenPair}(), Vector{Int8}([]))
+    return Node(Edge(i, i), Array{UInt8, 1}(),Dict{Int8, Int8}(), 0, Set{ForbiddenPair}(), Vector{UInt8}([]))
 end
 
 function Node(root_edge::LightGraphs.SimpleGraphs.SimpleEdge{Int64}, base_graph::SimpleGraph)::Node
-    comp_assign = Vector{Int8}([i for i in 1:nv(base_graph)])
-    return Node(root_edge, Array{Int8, 1}(), Dict{Int8, Int8}(), 0, Set{ForbiddenPair}(), comp_assign)
+    comp_assign = Vector{UInt8}([i for i in 1:nv(base_graph)])
+    return Node(root_edge, Array{UInt8, 1}(), Dict{Int8, Int8}(), 0, Set{ForbiddenPair}(), comp_assign)
 end
 
 mutable struct ZDD
@@ -156,7 +156,7 @@ function make_new_node(g::SimpleGraph,
                        i::Int,
                        x::Int8,
                        d::Int,
-                       frontiers::Array{Set{Int8}, 1},
+                       frontiers::Array{Set{UInt8}, 1},
                        lower_bound::Int8,
                        upper_bound::Int8,
                        zero_terminal::Node,
@@ -221,7 +221,7 @@ function make_new_node(g::SimpleGraph,
 end
 
 
-function add_vertex_as_component!(n′::Node, vertex::Int64, prev_frontier::Set{Int8})
+function add_vertex_as_component!(n′::Node, vertex::Int64, prev_frontier::Set{UInt8})
     """ Add `u` or `v` or both to n`.comp if they are not in
         `prev_frontier`
     """
@@ -232,7 +232,7 @@ function add_vertex_as_component!(n′::Node, vertex::Int64, prev_frontier::Set{
     nothing
 end
 
-function replace_components_with_union!(node::Node, Cᵤ::Int8, Cᵥ::Int8, fp_container::Vector{ForbiddenPair})
+function replace_components_with_union!(node::Node, Cᵤ::UInt8, Cᵥ::UInt8, fp_container::Vector{ForbiddenPair})
     """
     update fps to replace the smaller component with the larger component
     (TODO: rename)
@@ -258,7 +258,7 @@ function replace_components_with_union!(node::Node, Cᵤ::Int8, Cᵥ::Int8, fp_c
 end
 
 
-function connect_components!(n::Node, Cᵤ::Int8, Cᵥ::Int8)
+function connect_components!(n::Node, Cᵤ::UInt8, Cᵥ::UInt8)
     """
     If the two components are different, remove the smaller component from n.comp, and update n.comp_assign. Adjust n.comp_weights to remove the smaller-indexed component and add its weight into the larger one.
     """
@@ -272,7 +272,7 @@ function connect_components!(n::Node, Cᵤ::Int8, Cᵥ::Int8)
     end
 end
 
-function components(u::Int64, v::Int64, node::Node)::Tuple{Int8, Int8}
+function components(u::Int64, v::Int64, node::Node)::Tuple{UInt8, UInt8}
     """ Returns Cᵤ and Cᵥ which are the sets in `components` that contain
         vertices `u` and `v` respectively.
     """
@@ -297,8 +297,8 @@ function compute_frontier(edges, i::Int)
     end
 end
 
-function compute_all_frontiers(g::SimpleGraph, g_edges)::Vector{Set{Int8}}
-    frontiers = Vector{Set{Int8}}([Set{Int8}() for a in 1:ne(g)+1])
+function compute_all_frontiers(g::SimpleGraph, g_edges)::Vector{Set{UInt8}}
+    frontiers = Vector{Set{UInt8}}([Set{UInt8}() for a in 1:ne(g)+1])
     for i ∈ 1:ne(g)+1
         frontiers[i] = compute_frontier(g_edges, i)
     end
@@ -306,10 +306,10 @@ function compute_all_frontiers(g::SimpleGraph, g_edges)::Vector{Set{Int8}}
 end
 
 
-function nodes_from_edges(edges)::Set{Int8}
+function nodes_from_edges(edges)::Set{UInt8}
     """
     """
-    nodes = Set{Int8}()
+    nodes = Set{UInt8}()
     for e in edges
         push!(nodes, e.src)
         push!(nodes, e.dst)
@@ -317,7 +317,7 @@ function nodes_from_edges(edges)::Set{Int8}
     return nodes
 end
 
-function remove_vertex_from_node_fps!(node::Node, vertex::Int8, fp_container::Vector{ForbiddenPair})
+function remove_vertex_from_node_fps!(node::Node, vertex::UInt8, fp_container::Vector{ForbiddenPair})
     """
     """
     vertex_comp = node.comp_assign[vertex]
@@ -333,7 +333,7 @@ function remove_vertex_from_node_fps!(node::Node, vertex::Int8, fp_container::Ve
     adjust_node!(node, vertex_comp, fp_container)
 end
 
-function adjust_node!(node::Node, vertex_comp::Int8, fp_container::Vector{ForbiddenPair})
+function adjust_node!(node::Node, vertex_comp::UInt8, fp_container::Vector{ForbiddenPair})
     """
     """
     if vertex_comp in node.comp_assign
